fix: thinking graph UX — prevent node overlap, auto-fitView, extract SSE handlers
- Increase layout spacing: layer-0 radius 60→200+, layerRadius 180→350/400,
  chargeStrength -800→-1500, collideRadius 130→200
- Add 3-line text truncation to StreamingNode with WebkitLineClamp
- Compact node sizing: maxWidth 200→180, fontSize 12→11, tighter padding
- Remove misleading ring guide SVG overlay
- Remove nodeFloat margin-top animation that caused layout jank
- Auto-fitView when first SSE nodes arrive and on session completion
- Fix React Flow "parent needs width/height" warning with explicit h-full w-full
- Extract SSE mutation handlers to useSSEHandlers hook (ThinkingView 639→454 lines)
