+++ frontend/src/components/ThinkingView.tsx
@@ -1,6 +1,6 @@
 /**
- * ThinkingView — the concentric ring graph visualization for thinking sessions.
- * Shows the DAG with ring guides per layer, node detail panel, and controls.
+ * ThinkingView — concentric ring graph visualization for thinking sessions.
+ * SSE handlers extracted to useSSEHandlers hook.
  */
 import { useCallback, useEffect, useRef, useState } from "react";
 import {
@@ -15,12 +15,7 @@ import {
 } from "@xyflow/react";
 import "@xyflow/react/dist/style.css";
 import type { ThinkingSession, ThinkingNode } from "../types/thinking";
-import {
-  buildThinkingGraph,
-  nodeStyle,
-  concentricPosition,
-  LAYER_COLORS,
-} from "../lib/thinking-graph-builder";
+import { buildThinkingGraph, nodeStyle } from "../lib/thinking-graph-builder";
 import {
   LayerSpectrum,
   OpportunitiesBanner,
@@ -32,14 +27,7 @@ import { AgentFace } from "./AgentFace";
 import { usePathHighlight } from "../hooks/usePathHighlight";
 import { StreamingNode } from "./StreamingNode";
 import { useSSESession } from "../hooks/useSSESession";
-import type {
-  SSENodeStart,
-  SSENodeText,
-  SSENodeComplete,
-  SSELayerComplete,
-  SSESessionComplete,
-  SSEEdgesPayload,
-} from "../types/thinking";
+import { useSSEHandlers } from "../hooks/useSSEHandlers";
 
 const log = createLogger("thinking-view");
 
@@ -67,6 +55,22 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     handleNodeMouseEnter,
     handleNodeMouseLeave,
   } = usePathHighlight({ session, setNodes, setEdges });
+
+  const {
+    handleNodeStart,
+    handleNodeText,
+    handleNodeComplete,
+    handleEdges,
+    handleLayerComplete,
+    handleSessionComplete,
+  } = useSSEHandlers({
+    setNodes,
+    setEdges,
+    setSession,
+    setIsPending,
+    layerNodeCounts,
+    positionMap,
+  });
 
   const savePositions = useCallback((rfNodes: RFNode[]) => {
     for (const n of rfNodes) {
@@ -184,200 +188,6 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     }
   }, [sessionId, setNodes, setEdges, savePositions]);
 
-  // --- SSE streaming handlers ---
-
-  const handleNodeStart = useCallback(
-    (data: SSENodeStart) => {
-      const { id, layer, type, parent_ids } = data;
-      const count = (layerNodeCounts.current.get(layer) ?? 0) + 1;
-      layerNodeCounts.current.set(layer, count);
-      const idx = count - 1;
-
-      // Re-space existing nodes on this layer
-      setNodes((prev) => {
-        const updated = prev.map((n) => {
-          if (n.data.layer !== layer) return n;
-          const existingIdx = prev
-            .filter((p) => p.data.layer === layer)
-            .indexOf(n);
-          const pos = concentricPosition(layer, existingIdx, count);
-          positionMap.current.set(n.id, pos);
-          return { ...n, position: pos };
-        });
-
-        // Add the new streaming node
-        const pos = concentricPosition(layer, idx, count);
-        positionMap.current.set(id, pos);
-        const newNode: RFNode = {
-          id,
-          type: "streaming",
-          position: pos,
-          data: {
-            label: "",
-            type,
-            layer,
-            selected: true,
-            reasoning: "",
-            streaming: true,
-            parent_ids,
-          },
-          style: nodeStyle(type, true, layer),
-        };
-        return [...updated, newNode];
-      });
-
-      // Also track in session
-      setSession((prev) => {
-        if (!prev) return prev;
-        const newThinkingNode: ThinkingNode = {
-          id,
-          layer,
-          type: type as ThinkingNode["type"],
-          content: "",
-          reasoning: "",
-          sources: [],
-          parents: parent_ids,
-          selected: true,
-          metadata: {},
-        };
-        return {
-          ...prev,
-          nodes: [...prev.nodes, newThinkingNode],
-          current_layer: Math.max(prev.current_layer, layer),
-        };
-      });
-    },
-    [setNodes],
-  );
-
-  const handleNodeText = useCallback(
-    (data: SSENodeText) => {
-      const { id, field, delta } = data;
-      setNodes((prev) =>
-        prev.map((n) => {
-          if (n.id !== id) return n;
-          if (field === "content") {
-            return {
-              ...n,
-              data: { ...n.data, label: (n.data.label ?? "") + delta },
-            };
-          }
-          return {
-            ...n,
-            data: { ...n.data, reasoning: (n.data.reasoning ?? "") + delta },
-          };
-        }),
-      );
-
-      // Keep session in sync
-      setSession((prev) => {
-        if (!prev) return prev;
-        return {
-          ...prev,
-          nodes: prev.nodes.map((n) => {
-            if (n.id !== id) return n;
-            if (field === "content")
-              return { ...n, content: n.content + delta };
-            return { ...n, reasoning: n.reasoning + delta };
-          }),
-        };
-      });
-    },
-    [setNodes],
-  );
-
-  const handleNodeComplete = useCallback(
-    (data: SSENodeComplete) => {
-      const { id, confidence, metadata } = data;
-      setNodes((prev) =>
-        prev.map((n) => {
-          if (n.id !== id) return n;
-          return {
-            ...n,
-            data: { ...n.data, streaming: false, confidence },
-          };
-        }),
-      );
-
-      setSession((prev) => {
-        if (!prev) return prev;
-        return {
-          ...prev,
-          nodes: prev.nodes.map((n) =>
-            n.id !== id
-              ? n
-              : { ...n, metadata: { ...n.metadata, ...metadata, confidence } },
-          ),
-        };
-      });
-    },
-    [setNodes],
-  );
-
-  const handleEdges = useCallback(
-    (payload: SSEEdgesPayload) => {
-      const data = payload.edges;
-      const newEdges: RFEdge[] = data.map((e, i) => {
-        const isMatch = e.relationship === "matches";
-        const isConfirm =
-          e.relationship === "compounds" || e.relationship === "causes";
-        return {
-          id: `sse-e-${Date.now()}-${i}-${e.source}-${e.target}`,
-          source: e.source,
-          target: e.target,
-          label: e.relationship,
-          type: "smoothstep",
-          animated: true,
-          markerEnd: {
-            type: "arrowclosed" as const,
-            color: isMatch
-              ? "rgba(34, 197, 94, 0.6)"
-              : "rgba(255, 255, 255, 0.3)",
-            width: 15,
-            height: 15,
-          },
-          style: {
-            stroke: isMatch
-              ? "rgba(34, 197, 94, 0.4)"
-              : isConfirm
-                ? "rgba(255, 255, 255, 0.2)"
-                : "rgba(107, 115, 148, 0.3)",
-            strokeWidth: isMatch ? 2 : 1.5,
-          },
-          labelStyle: {
-            fill: isMatch ? "#86efac" : "#6b7394",
-            fontSize: 10,
-          },
-        };
-      });
-      setEdges((prev) => [...prev, ...newEdges]);
-
-      // Keep session edges in sync
-      setSession((prev) => {
-        if (!prev) return prev;
-        return { ...prev, edges: [...prev.edges, ...data] };
-      });
-    },
-    [setEdges],
-  );
-
-  const handleLayerComplete = useCallback((data: SSELayerComplete) => {
-    setSession((prev) => {
-      if (!prev) return prev;
-      return { ...prev, current_layer: data.layer };
-    });
-    log.info("Layer", data.layer, "complete:", data.controller.summary);
-  }, []);
-
-  const handleSessionComplete = useCallback((data: SSESessionComplete) => {
-    setSession((prev) => {
-      if (!prev) return prev;
-      return { ...prev, status: data.status as ThinkingSession["status"] };
-    });
-    setIsPending(false);
-    log.info("Session complete:", data.status);
-  }, []);
-
   // Wire SSE hook — active only during "thinking" status
   const { connected } = useSSESession(
     session?.status === "thinking" ? sessionId : null,
@@ -395,6 +205,26 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
 
   const isThinking = session?.status === "thinking" || isPending;
 
+  // Auto-fitView when nodes arrive during SSE streaming or session completes
+  const prevNodeCount = useRef(0);
+  useEffect(() => {
+    const count = nodes.length;
+    // Fit on first batch of nodes and when session completes
+    if (
+      count > 0 &&
+      (prevNodeCount.current === 0 ||
+        session?.status === "complete" ||
+        session?.status === "paused")
+    ) {
+      const delay = prevNodeCount.current === 0 ? 300 : 100;
+      setTimeout(
+        () => rfInstance.current?.fitView({ padding: 0.15, duration: 500 }),
+        delay,
+      );
+    }
+    prevNodeCount.current = count;
+  }, [nodes.length, session?.status]);
+
   // Polling fallback — only when SSE is not connected
   useEffect(() => {
     if (connected || session?.status !== "thinking") return;
@@ -402,11 +232,9 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     return () => clearInterval(interval);
   }, [connected, session?.status, loadSessionIncremental]);
 
-  // Step to next layer
   const handleStep = useCallback(async () => {
     if (isThinking || !session) return;
     setIsPending(true);
-    log.info("Stepping to layer", session.current_layer + 1);
     try {
       await graphApi.thinkStep(sessionId);
       await loadSessionIncremental();
@@ -416,11 +244,9 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     setIsPending(false);
   }, [sessionId, session, isThinking, loadSessionIncremental]);
 
-  // Match against value pool
   const handleMatch = useCallback(async () => {
     if (isThinking || !session) return;
     setIsPending(true);
-    log.info("Matching against value pool");
     try {
       await graphApi.matchValues(sessionId);
       await loadSessionIncremental();
@@ -430,18 +256,14 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     setIsPending(false);
   }, [sessionId, session, isThinking, loadSessionIncremental]);
 
-  // Pane click: useCallback to avoid stale closure in React Flow
   const handlePaneClick = useCallback(() => {
     unpinPath();
     setSelectedNode(null);
   }, [unpinPath]);
 
-  // Fallback: container div click also dismisses if target is the pane itself
-  // (not a node or edge inside the pane — those are handled by onNodeClick)
   const handleContainerClick = useCallback(
     (e: React.MouseEvent<HTMLDivElement>) => {
       const target = e.target as HTMLElement;
-      // Only fire if clicking directly on the pane background, not on nodes/edges/UI
       if (
         target.classList.contains("react-flow__pane") ||
         target.classList.contains("react-flow__renderer")
@@ -453,7 +275,6 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     [unpinPath],
   );
 
-  // Escape key also clears selectedNode (hook handles unpinPath)
   useEffect(() => {
     const handleKeyDown = (e: KeyboardEvent) => {
       if (e.key === "Escape") setSelectedNode(null);
@@ -462,7 +283,6 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     return () => window.removeEventListener("keydown", handleKeyDown);
   }, []);
 
-  // Click node: show detail panel. Pin path only for opportunity nodes.
   const handleNodeClick = useCallback(
     async (_: React.MouseEvent, node: RFNode) => {
       if (!session) return;
@@ -480,7 +300,6 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     [session, pinnedNodeId, unpinPath, pinPath],
   );
 
-  // Toggle node via detail panel
   const handleToggle = useCallback(
     async (nodeId: string, selected: boolean) => {
       try {
@@ -499,39 +318,21 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
     (session?.current_layer ?? 0) < (session?.max_depth ?? 3);
   const canMatch =
     session?.status === "paused" && (session?.current_layer ?? 0) > 0;
-  const maxLayer = Math.max(0, ...(session?.nodes.map((n) => n.layer) ?? [0]));
+  const hasEffectNodes = nodes.some(
+    (n) =>
+      (n.data as Record<string, unknown>).layer !== undefined &&
+      ((n.data as Record<string, unknown>).layer as number) > 0,
+  );
 
   return (
-    <div className="flex-1 relative" onClick={handleContainerClick}>
-      {/* Ring guides — colored concentric circles per layer */}
-      <svg
-        className="absolute inset-0 w-full h-full pointer-events-none z-0"
-        style={{ opacity: 0.2 }}
-      >
-        {Array.from({ length: maxLayer + 2 }, (_, i) => {
-          const r = i * 140;
-          const layerColor = LAYER_COLORS[Math.min(i, LAYER_COLORS.length - 1)];
-          return r > 0 ? (
-            <circle
-              key={i}
-              cx="50%"
-              cy="50%"
-              r={r}
-              fill="none"
-              stroke={layerColor}
-              strokeWidth={1}
-              strokeDasharray="4 8"
-              opacity={0.4}
-            />
-          ) : null;
-        })}
-      </svg>
-
+    <div
+      className="flex-1 relative h-full w-full"
+      onClick={handleContainerClick}
+    >
       <LayerSpectrum session={session} />
 
       {/* Controls bar */}
       <div className="absolute top-4 left-1/2 -translate-x-1/2 z-20 flex items-center gap-3">
-        {/* Status */}
         <div className="glass-card px-4 py-2 flex items-center gap-3">
           <span className="text-xs text-text-muted uppercase tracking-widest">
             Layer {session?.current_layer ?? 0}/{session?.max_depth ?? 3}
@@ -557,7 +358,6 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
           )}
         </div>
 
-        {/* Action buttons */}
         {canStep && (
           <button
             onClick={handleStep}
@@ -603,21 +403,15 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
       <OpportunitiesBanner session={session} />
 
       {/* Agent face: only while thinking AND before any effect nodes arrive */}
-      {isThinking &&
-        !nodes.some(
-          (n) =>
-            (n.data as Record<string, unknown>).layer !== undefined &&
-            ((n.data as Record<string, unknown>).layer as number) > 0,
-        ) && (
-          <div
-            className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 z-10 pointer-events-none"
-            style={{ opacity: 0.4 }}
-          >
-            <AgentFace size={80} />
-          </div>
-        )}
+      {isThinking && !hasEffectNodes && (
+        <div
+          className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 z-10 pointer-events-none"
+          style={{ opacity: 0.4 }}
+        >
+          <AgentFace size={80} />
+        </div>
+      )}
 
-      {/* React Flow graph */}
       <ReactFlow
         nodeTypes={nodeTypes}
         nodes={nodes}
@@ -634,8 +428,8 @@ export function ThinkingView({ sessionId, onReset }: ThinkingViewProps) {
           rfInstance.current = instance;
         }}
         fitView
-        fitViewOptions={{ padding: 0.1 }}
-        minZoom={0.2}
+        fitViewOptions={{ padding: 0.15 }}
+        minZoom={0.15}
         maxZoom={2}
         proOptions={{ hideAttribution: true }}
       >
